Inline __get__ and __set__ definitions in injection script
Without defining __get__ and __set__, most javascript runtimes will
throw an error when in strict mode, complaining about a lack of definition
for those two variables. This happens since those function declarations
are `eval`d and therefore the runtime has no way of knowing about them.

To fix this, we can simply call toString() on __get__ and __set__ as the
RHS of the `module.exports.__get__` and `module.exports.__set__`
assignments, respectively. This will allow this transform to work when
modules have strict mode enabled.